Update App Icons and Dependencies
Updated the Android and iOS app icons to use a new design. Added new icon assets for various resolutions and updated the AndroidManifest.xml to reference the new launcher icon. Additionally, added a new color resource for the launcher background in Android.

Updated the pubspec.yaml to include the flutter_launcher_icons package for generating app icons and configured it to use the new icon assets. Also, updated the pubspec.lock to reflect the new dependencies and their versions.

This commit includes:
- New icon assets for Android and iOS.
- Updated AndroidManifest.xml to reference the new launcher icon.
- Added a new color resource for the Android launcher background.
- Updated pubspec.yaml and pubspec.lock to include the flutter_launcher_icons package and its dependencies.