Add dots_indicator package to pubspec.yaml
This commit introduces the `dots_indicator` package to the project, version 3.0.0. The package is added as a direct main dependency in the `pubspec.yaml` file, and the corresponding entry is updated in the `pubspec.lock` file to reflect the new dependency. The `dots_indicator` package is used to provide a simple and customizable way to display pagination dots in Flutter applications.