Update Firebase Hosting Cache and Asset Files
Updated the Firebase Hosting cache file with new entries for various assets, including icons, splash images, and fonts. This includes new versions of `manifest.json`, `favicon.png`, and multiple `splash/img` and `icons` files. Additionally, updated the `version.json`, `index.html`, `flutter_service_worker.js`, and other asset files to reflect the latest changes in the project. This commit ensures that the hosting cache is up-to-date with the latest assets and configurations.