Add OrDivider widget to login view
This commit introduces the `OrDivider` widget, which is used to separate sections in the login view. The widget consists of a row containing two horizontal dividers with the text "او" (meaning "or" in Arabic) centered between them. The dividers are styled with a light gray color and adjusted for responsiveness using `flutter_screenutil`. The text is styled using the `AppTextStyles.bodyBasaSemiBold16` from the app's theme, with the color set to black.