Add Login View and Related Widgets
This commit introduces the LoginView and its associated widgets, including the LoginViewBody, DontHaveAnAccont, SocialLoginButton, and OrDivider. The LoginView is a stateless widget that sets up the Scaffold for the login screen, including a custom app bar and a SafeArea containing the LoginViewBody. The LoginViewBody is a stateful widget that manages the login form, including email and password fields, a "Forgot Password" link, a login button, and social login options (Google, Apple, Facebook). The DontHaveAnAccont widget provides a link for users without an account to create one, and the SocialLoginButton widget is used to display buttons for social login options.