Add AuthRepo interface for user authentication
This commit introduces the `AuthRepo` interface within the `auth` feature's domain layer. The interface defines a method `createUserwithEmailandPassword` that returns a `Future` of `Either<Faliure, UserEntity>`. This method is responsible for handling user creation with email and password, encapsulating the potential failure or successful creation of a `UserEntity`. The addition of this interface sets the foundation for implementing authentication logic in the application.