Add UserEntity class to represent user data in the domain layer
This commit introduces the `UserEntity` class within the `lib/features/auth/domain/entites/` directory. The `UserEntity` class is designed to encapsulate user-related data, including `name`, `email`, and `uId`. This entity will serve as a foundational component for handling user information within the domain layer of the authentication feature.

The class is implemented with a constructor that requires all three fields to be provided, ensuring that instances of `UserEntity` are always fully populated. This design choice helps maintain data integrity and consistency throughout the application.