Add functionality to check if user exists in Firestore before adding
Previously, the application would always attempt to add a new user to Firestore upon authentication, regardless of whether the user already existed. This commit introduces a new method in the `FirestoreService` class to check if a document exists in Firestore. The `AuthRepoImpl` class is updated to use this method to first check if the user exists before attempting to add or retrieve the user data. This ensures that duplicate user entries are avoided and existing user data is correctly retrieved.

- Added `checkIfDataExists` method to `FirestoreService` to check if a document exists in Firestore.
- Updated `AuthRepoImpl` to use `checkIfDataExists` before adding or retrieving user data.
- Added `isUerExists` endpoint in `EndPoint` class for the Firestore path.