Update Firebase Hosting Cache and Configuration Files
Additionally, restructured the `firebase.json` file to improve readability and maintainability. The configuration now follows a more standard JSON format with proper indentation and structure.

In the `login_view_body.dart` file, conditional rendering of the Apple Sign-In button was refactored to use `kIsWeb` and `Platform.isIOS` checks. This ensures that the button is only rendered on iOS devices and not on the web, improving performance and reducing unnecessary code.