Add HomeView and HomeViewBody Widgets
Created the `HomeView` and `HomeViewBody` widgets in the `lib/features/home/presentation/view` directory. The `HomeView` widget serves as the main view for the home screen, utilizing a `Scaffold` with a `SafeArea` to ensure content is displayed correctly on different devices. The `HomeViewBody` widget is a stateless widget that contains a `CustomScrollView` with a `SliverToBoxAdapter` to hold the `CutsomHomeAppBar` and `CustomSearch` widgets, ensuring a responsive layout.