Add GoRouter configuration for dashboard navigation
Created a new GoRouter configuration to handle navigation within the dashboard section of the app. This includes a new `router_d_b.dart` file that defines the router setup and a `router_name_d_b.dart` file that contains route names. The `main_dashs_board.dart` file has been updated to initialize the router and integrate it with the MaterialApp. This change also includes initializing Firebase and setting up a custom Bloc observer.